Refactor project components and styles
Why:
The general structure of components and styles was not well defined and
glue together by code definitions.

How:
I try to keep the code DRY and reducing code that doesn't make sense,
getting rid of it on components and sass style. As well, we defined
something more modular, focus on reusability.

Tags: refactor